refactored feature files upto and including print_files
